refactor: #24 status 패키지 리팩토링
