+++ logistics/src/main/java/msa/logistics/service/logistics/delivery/domain/DeliveryStatus.java
@@ -1,4 +1,4 @@
-package msa.logistics.service.logistics.delivery.dto;
+package msa.logistics.service.logistics.delivery.domain;
 
 public enum DeliveryStatus {
     WAITING, IN_TRANSIT, ARRIVED, DELIVERED
