+++ logistics/src/main/java/msa/logistics/service/logistics/delivery/domain/RouteStatus.java
@@ -1,4 +1,4 @@
-package msa.logistics.service.logistics.delivery.dto;
+package msa.logistics.service.logistics.delivery.domain;
