+++ logistics/src/main/java/msa/logistics/service/logistics/delivery/dto/DeliveryRouteEditRequestDto.java
@@ -1,6 +1,7 @@
 package msa.logistics.service.logistics.delivery.dto;
 
 import lombok.*;
+import msa.logistics.service.logistics.delivery.domain.RouteStatus;
 
 @AllArgsConstructor
 @NoArgsConstructor
